backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
openai
supabase>=2.0.0
python-pptx
python-docx
PyJWT
bcrypt
qianfan
PyPDF2
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
unstructured[pdf,docx,pptx]
dashscope
pypdf
